+++ .gitignore
@@ -35,4 +35,4 @@ yarn-error.log*
 *.tsbuildinfo
 next-env.d.ts
 
-/public/notion-static
+public/notion-static
