+++ build-scripts/prepare-static/constants.js
@@ -1,0 +1,9 @@
+export const VERCEL_PUBLIC_PATH = '/notion-static/';
+
+export const VERCEL_OUTPUT_PATH = `./public${VERCEL_PUBLIC_PATH}`;
+
+export const VERCEL_PUBLIC_IMAGES_PATH = `${VERCEL_PUBLIC_PATH}images/`;
+
+export const IMAGES_URLS_PATH = `${VERCEL_OUTPUT_PATH}images/`;
+
+export const PLACEMARKS_CACHE_PATH = `${VERCEL_OUTPUT_PATH}placemarks.json`;
