+++ package.json
@@ -7,9 +7,9 @@
   "type": "module",
   "scripts": {
     "dev": "next dev",
-    "build": "npm run preload-static && next build",
+    "build": "next build",
     "start": "next start",
-    "preload-static": "node ./build-scripts/cache-images.js",
+    "preload-static": "node ./build-scripts/prepare-static/index.js",
     "lint": "next lint",
     "postinstall": "pnpm preload-static",
     "devPreinstall": "husky install"
